14-30/11

+++ admin/ajax/features_facilities.php
@@ -39,11 +39,17 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_feature']];
 
-        // $check_q = 
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?', [$frm_data['rem_feature']], 'i');
         
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q, $values,'i');
-        echo $res;  
+        if(mysqli_num_rows($check_q)==0)
+        {
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q, $values,'i');
+            echo $res;  
+        }
+        else{
+            echo 'room_added';
+        }
 
     }
 
@@ -84,11 +90,26 @@
     {
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_facility']];
-        
-        $q = "DELETE FROM `facilities` WHERE `id`=?";
-        $res = delete($q,$values,'i');
-        echo $res;  
 
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?', [$frm_data['rem_facility']], 'i');
+
+        if(mysqli_num_rows($check_q)==0){
+            $pre_q ="SELECT * FROM `facilities` WHERE `id`=?";
+            $res = select($pre_q, $values, 'i');
+            $img = mysqli_fetch_assoc($res);
+
+            if(deleteImage($img['icon'], FACILITIES_FOLDER)){
+                $q = "DELETE FROM `facilities` WHERE `id`=?";
+                $res = delete($q,$values,'i');
+                echo $res; 
+            } 
+            else{
+                echo 0;
+            }
+        }
+        else{
+            echo 'room_added';
+        } 
     }
